Tests for queries in client.jsx
